requirement.txt
flask>=2.0.1
openai>=1.0.0
python-dotenv>=0.19.0
moviepy>=1.0.3
pydantic>=1.8.2
pydantic-settings>=2.0.3
google-generativeai>=0.3.2
PyPDF2>=3.0.1